hack: use "set -e" to fail scripts on error
It's error prone when scripts don't check for the success of their
commands. Don't do that, use `set -e`.

If you find that a step in the scripts can fail for valid reasons and
you don't want that to abort the script, then catch the exit code (for
example with `|| true`).