hack: detect the DPU kind for e2e host
Currently, for e2e test we only have the config files
"config-dpu-{,host}.yaml". Those currently only work with IPU systems.

Next, we will add "config-marvell-dpu-{,host}.yaml". Those files are
currently different than the ones for IPU. Hence, the "cda_deploy.sh"
needs to detect which one to use.

Add code to detect this based on the current host name.

In the future, we may want to unify all DPU configuration files, and
fill them with information from the google spread sheet. Note that
currently does not work, because the files are different and the right
information cannot be fetched from the spread sheet. But even if that is
implemented, and cda-deploy fills in the information from the spread
sheet, the detection is *also* based on the hostname. The only
difference will be, that it's then all encoded in the spread sheet. This
will happen later.

For now, add code to detect the DPU kind and select the right config
file.