taskfile: add "redeploy" and "run-e2e-test-suite" taskfile targets
In the past, CDA deployed the dpu-operator via a "post_config:
dpu_operator_{host,dpu}" settings. This meant for example, that a `make
deploy_clusters` would have also deployed the dpu-operator.

This is no longer the case. Now, the dpu-operator is installed during
`make e2e-test`, but there isn't a separate target for installing the
dpu-operator At most, you could redo that by running the existing
subtasks "build-image-all", "undeploy" and "deploy" together.

Add a "redeploy" target, which installs the dpu-operator.

Note that contrary to the general name, "redeploy" is not used to
install dpu-operator in a certain cluster (e.g. by providing a
KUBECONFIG). Instead, it deploys two clusters (hostside and dpuside)
with hard-coded kubeconfig files. That is in line with what "deploy"
and "undeploy" targets do.

Likewise, in the past we could run `make e2e-test-suite`, to only run
the tests from "./e2e-tests". This too no longer works. Re-add a
"run-e2e-test-suite" target for that.